breaks components out of Room into container components
forgoes using disclosureMachine for now
