+++ src/components/Room.tsx
@@ -2,53 +2,21 @@ import React, { useCallback, useEffect } from "react"
 import { useSelector } from "@xstate/react"
 import Konami from "react-konami-code"
 
-import { Box, Grid, GridItem, Hide, Show, useToken } from "@chakra-ui/react"
+import { Box, Grid, GridItem, Show, useToken } from "@chakra-ui/react"
 
-import FormAdminMeta from "./FormAdminMeta"
-import FormAdminArtwork from "./FormAdminArtwork"
-import FormAdminSettings from "./FormAdminSettings"
-import FormAddToQueue from "./FormAddToQueue"
 import PlayerUi from "./PlayerUi"
-import FormUsername from "./FormUsername"
-import FormPassword from "./FormPassword"
 import Chat from "./Chat"
-import UserList from "./UserList"
-import Modal from "./Modal"
-import Drawer from "./Drawer"
 import Sidebar from "./Sidebar"
-import AboutContent from "./AboutContent"
-import socket from "../lib/socket"
+import Overlays from "./Overlays"
 import useGlobalContext from "./useGlobalContext"
-import FormTheme from "./FormTheme"
-import DrawerBookmarks from "./DrawerBookmarks"
 import KeyboardShortcuts from "./KeyboardShortcuts"
-import DrawerPlaylist from "./DrawerPlaylist"
 
 const isEditingSelector = (state) =>
   state.matches("connected.participating.editing")
-const playlistActiveSelector = (state) => state.matches("active")
-const isEditingUsernameSelector = (state) =>
-  state.matches("connected.participating.editing.username")
-const isModalViewingListenersSelector = (state) =>
-  state.matches("connected.participating.modalViewing.listeners")
-const isModalViewingHelpSelector = (state) =>
-  state.matches("connected.participating.modalViewing.help")
-const isEditingMetaSelector = (state) =>
-  state.matches("connected.participating.editing.meta")
-const isEditingArtworkSelector = (state) =>
-  state.matches("connected.participating.editing.artwork")
-const isEditingSettingsSelector = (state) =>
-  state.matches("connected.participating.editing.settings")
-const isEditingPreferencesSelector = (state) =>
-  state.matches("connected.participating.editing.preferences")
-const isAddingToQueueSelector = (state) =>
-  state.matches("connected.participating.editing.queue")
-const playlistSelector = (state) => state.context.playlist
+
 const isAdminSelector = (state) => state.context.isAdmin
 const isNewUserSelector = (state) => state.context.isNewUser
-const isUnauthorizedSelector = (state) => state.matches("unauthorized")
-const currentUserSelector = (state) => state.context.currentUser
-const passwordErrorSelector = (state) => state.context.passwordError
+const playlistSelector = (state) => state.context.playlist
 const listenersSelector = (state) => state.context.listeners
 
 const Room = () => {
@@ -56,58 +24,9 @@ const Room = () => {
 
   const globalServices = useGlobalContext()
   const isEditing = useSelector(globalServices.roomService, isEditingSelector)
-  const playlistActive = useSelector(
-    globalServices.playlistService,
-    playlistActiveSelector,
-  )
-  const isEditingUsername = useSelector(
-    globalServices.roomService,
-    isEditingUsernameSelector,
-  )
-  const isModalViewingListeners = useSelector(
-    globalServices.roomService,
-    isModalViewingListenersSelector,
-  )
-  const isModalViewingHelp = useSelector(
-    globalServices.roomService,
-    isModalViewingHelpSelector,
-  )
-  const isEditingMeta = useSelector(
-    globalServices.roomService,
-    isEditingMetaSelector,
-  )
-  const isEditingArtwork = useSelector(
-    globalServices.roomService,
-    isEditingArtworkSelector,
-  )
-  const isEditingSettings = useSelector(
-    globalServices.roomService,
-    isEditingSettingsSelector,
-  )
-  const isEditingPreferences = useSelector(
-    globalServices.roomService,
-    isEditingPreferencesSelector,
-  )
 
-  const isAddingToQueue = useSelector(
-    globalServices.roomService,
-    isAddingToQueueSelector,
-  )
   const isNewUser = useSelector(globalServices.authService, isNewUserSelector)
   const isAdmin = useSelector(globalServices.authService, isAdminSelector)
-  const isUnauthorized = useSelector(
-    globalServices.authService,
-    isUnauthorizedSelector,
-  )
-  const currentUser = useSelector(
-    globalServices.authService,
-    currentUserSelector,
-  )
-
-  const passwordError = useSelector(
-    globalServices.authService,
-    passwordErrorSelector,
-  )
 
   const playlist = useSelector(globalServices.playlistService, playlistSelector)
 
@@ -125,32 +44,12 @@ const Room = () => {
     }
   }, [isAdmin])
 
-  const hideListeners = useCallback(
-    () => globalServices.roomService.send("CLOSE_VIEWING"),
-    [globalServices],
-  )
-
-  const hideEditForm = useCallback(
-    () => globalServices.roomService.send("CLOSE_EDIT"),
-    [globalServices],
-  )
-
   const onOpenReactionPicker = useCallback(
     (reactTo: any) => {
       globalServices.roomService.send("TOGGLE_REACTION_PICKER", {
         reactTo,
       })
     },
-    [globalServices.roomService],
-  )
-
-  const handleEditSettings = useCallback(
-    () => globalServices.roomService.send("ADMIN_EDIT_SETTINGS"),
-    [globalServices.roomService],
-  )
-
-  const handleEditUser = useCallback(
-    () => globalServices.roomService.send("EDIT_USERNAME"),
     [globalServices.roomService],
   )
 
@@ -224,103 +123,7 @@ const Room = () => {
         </GridItem>
       </Grid>
 
-      <DrawerPlaylist isOpen={playlistActive} />
-
-      <Hide above="sm">
-        <Drawer
-          isOpen={isModalViewingListeners}
-          isFullHeight
-          heading={`Listeners (${listeners.length})`}
-          size={["sm", "lg"]}
-          onClose={() => hideListeners()}
-        >
-          <Box p="sm" overflow="auto" h="100%">
-            <div>
-              <UserList
-                showHeading={false}
-                onEditSettings={handleEditSettings}
-                onEditUser={handleEditUser}
-              />
-            </div>
-          </Box>
-        </Drawer>
-      </Hide>
-
-      {isAdmin && <DrawerBookmarks />}
-
-      <FormUsername
-        isOpen={isEditingUsername}
-        currentUser={currentUser}
-        onClose={hideEditForm}
-        onSubmit={(username) => {
-          globalServices.authService.send({
-            type: "UPDATE_USERNAME",
-            data: username,
-          })
-          hideEditForm()
-        }}
-      />
-
-      <FormPassword
-        isOpen={isUnauthorized}
-        error={passwordError}
-        onSubmit={(password) => {
-          globalServices.authService.send({
-            type: "SET_PASSWORD",
-            data: password,
-          })
-        }}
-      />
-
-      <Modal
-        isOpen={isModalViewingHelp}
-        onClose={() => globalServices.roomService.send("CLOSE_VIEWING")}
-        heading="???"
-      >
-        <AboutContent />
-      </Modal>
-
-      <Modal
-        isOpen={isEditingPreferences}
-        onClose={hideEditForm}
-        heading="Preferences"
-      >
-        <FormTheme />
-      </Modal>
-
-      <Modal
-        isOpen={isEditingMeta}
-        onClose={hideEditForm}
-        heading="Set Station Info"
-      >
-        <FormAdminMeta onSubmit={(value) => socket.emit("fix meta", value)} />
-      </Modal>
-
-      <Modal
-        isOpen={isEditingArtwork}
-        onClose={hideEditForm}
-        heading="Set Cover Artwork"
-      >
-        <FormAdminArtwork
-          onSubmit={(value) => socket.emit("set cover", value)}
-        />
-      </Modal>
-
-      <Modal
-        isOpen={isAddingToQueue}
-        onClose={hideEditForm}
-        heading="Add to play queue"
-      >
-        <FormAddToQueue onSubmit={(value) => socket.emit("set cover", value)} />
-      </Modal>
-
-      <FormAdminSettings
-        isOpen={isEditingSettings}
-        onClose={hideEditForm}
-        onSubmit={(value) => {
-          socket.emit("settings", value)
-        }}
-      />
+      <Overlays />
     </Box>
   )
 }
